requirements.txt
asyncio~=3.4.3
LayerAkiraSDK~=1.0.0a94
starknet-py~=0.25.0